Standardise sample counts to be `long`

+++ include/plugin.h
@@ -26,7 +26,7 @@ struct PluginState {
   ~PluginState();
 
   DB_fileinfo_t fFileInfo;
-  int readsample;
+  long readsample;
   bool fInit;
   uint32_t hints;
   TrackMetadata fMetadata;
